for that person who commented on my youtube video
this is everything i have (right now). i kind of stopped working on this project around the time the series stopped because i got too busy with college work (still am lol). hopefully i get back to it but this should be enough if you want to play around with it.